requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
openai>=1.6.1
groq>=0.4.1
beautifulsoup4>=4.12.2
requests>=2.31.0
selenium>=4.15.0
webdriver-manager>=4.0.1
undetected-chromedriver>=3.5.4
numpy>=1.26.2
scikit-learn>=1.3.2
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
aiohttp>=3.9.1
linkedin-api>=2.2.0
redis>=5.0.1
celery>=5.3.4
psycopg2-binary>=2.9.9
pyyaml>=6.0.1
sentence-transformers>=2.2.2
